add confident score

+++ train_val.py
@@ -166,7 +166,7 @@ def train(model, body_part, framework, sample_fps, video_len=99999, ori_videos=F
                 inputs = inputs.to(dtype=dtype, device=device)
             elif 'gcn_' in model:
                 inputs, (int_labels, att_labels, act_labels) = data
-            int_labels, att_labels, act_labels = int_labels.to(dtype=dtype, device=device), att_labels.to(
+            int_labels, att_labels, act_labels = int_labels.to(dtype=torch.long, device=device), att_labels.to(
                 device), act_labels.to(device)
             if framework == 'intention':
                 int_outputs = net(inputs)
